Fix race condition again

+++ denorm/denorms.py
@@ -831,16 +831,19 @@ def flush_single(content_type_id, object_id, content_type=None):
             return
 
         klass = content_type.model_class()
-        try:
-            obj = klass.objects.select_for_update(of=("self",), skip_locked=True).get(
-                pk=object_id
-            )
-        except klass.DoesNotExist:
-            res.delete()
-            return
 
-        if obj is None:
-            # nowait=True raises OperationalError when row is locked, skip this update
+        if klass.objects.filter(pk=object_id).exists():
+            try:
+                obj = klass.objects.select_for_update(
+                    of=("self",), skip_locked=True
+                ).get(pk=object_id)
+            except klass.DoesNotExist:
+                # Locked
+                return
+
+        else:
+            # Truly gone
+            res.delete()
             return
 
         func_names = set(list(res.values_list("func_name", flat=True)))
